refactor: 패키지 구조 변경 및 SecurityUtil 삭제
- 가독성을 위해 dto 패키지 구조를 세분화
- authenticationPrincipal 어노테이션을 사용하면 돼서 SecurityUtil 클래스 삭제
